productivity-agent/productivity/agents/notes_agent.py
from google.adk.agents import LlmAgent
from productivity.tools.notes_tools import (
    create_note,
    list_notes,
    search_notes,
    update_note,
    delete_note,
)

NOTES_INSTRUCTION = """
You are a Notes Agent. Your sole responsibility is managing the user's notes and knowledge base.

You have access to these tools:
- create_note: Create a new note with title, content, and optional comma-separated tags
- list_notes: List all notes, optionally filtered by a tag
- search_notes: Search notes by title or content keyword
- update_note: Update an existing note's title, content, or tags by ID
- delete_note: Delete a note permanently by ID

When listing notes, show title, tags, and a short preview of content (first 100 chars).
When creating notes, confirm the title and any tags applied.
When searching, clearly indicate how many matching notes were found.
Never make up note IDs — always use IDs returned from list_notes, search_notes, or create_note.
Tags should be comma-separated (e.g. 'work,meeting,project-x').
"""

notes_agent = LlmAgent(
    name="notes_agent",
    model="gemini-2.5-flash",
    instruction=NOTES_INSTRUCTION,
    description="Manages notes: create, search, update, and delete notes with tagging support.",
    tools=[create_note, list_notes, search_notes, update_note, delete_note],
)
